% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Y.ODEDiff.R
\name{ODEDiff}
\alias{ODEDiff}
\title{Ordinary Differential Equation Differentiation}
\usage{
ODEDiff(t, y, param)
}
\arguments{
\item{t}{时间，自变量}

\item{y}{各人群状态的初始值向量，依次为 S、E、I1、I2、R}

\item{param}{参数列表，包含 mu、lamda、beta、gamma1、gamma2 和 N}
}
\value{
返回一个列表，包含了 S、E、I1、I2 和 R 各人群状态的变化率
}
\description{
这个函数用于定义 SEIR 模型的常微分方程，
描述了在传染病传播过程中易感者（S）、暴露者（E）、首次感染者（I1）、
二次感染者（I2）和康复者（R）各类人群的人数变化。
}
\examples{
 \dontrun{
# 设置评估参数的初始值
times <- seq(0, 156, by = 1/7)
param <- c(mu = 0.000, lamda = 0.03, beta = 4, gamma1 = 0.1, gamma2 = 0.05, N = 1)
init <- c(S = 0.9999, E = 0.00008, I1 = 0.00001, I2 = 0, R = 0)

# 调用常微分方程求解函数
result <- ODEDiff(t = times, y = init, param = param)
result <- as.data.frame(result)

# 结果画图
seirplot <- ggplot(data = result, aes(x = time)) +
  geom_line(aes(y = S, color = "S"), linewidth = 1) +
  geom_line(aes(y = E, color = "E"), linewidth = 1) +
  geom_line(aes(y = I1, color = "I1"), linewidth = 1) +
  geom_line(aes(y = I2, color = "I2"), linewidth = 1) +
  geom_line(aes(y = R, color = "R"), linewidth = 1) +
  labs(x = "Time", y = "Ratio") +
  scale_color_manual(name = "SEIR",
                     values = c("S" = "orange", "E" = "purple", "I1" = "red", "I2" = "blue", "R" = "green")) +
  theme_minimal()

# 绘制仿真结果并保存为矢量文件
print(seirplot)
ggsave(filename = "Y.ODEDiff_plot.pdf", plot = seirplot, width = 7, height = 6)
ggsave(filename = "Y.ODEDiff_plot.svg", plot = seirplot, width = 7, height = 6)
}

}
